Fix se muestran los decimales en detalle del pedido

+++ services/sapadapter-service/Omicron.SapAdapter.DataAccess/DAO/Sap/SapDao.cs
@@ -144,7 +144,7 @@ namespace Omicron.SapAdapter.DataAccess.DAO.Sap
                                    CodigoProducto = d.ProductoId,
                                    DescripcionProducto = p.LargeDescription,
                                    DescripcionCorta = p.ProductoName,
-                                   QtyPlanned = (int)dp.Quantity,
+                                   QtyPlanned = dp.Quantity,
                                    QtyPlannedDetalle = (int)d.Quantity,
                                    FechaOf = dp.PostDate.HasValue ? dp.PostDate.Value.ToString("dd/MM/yyyy") : string.Empty,
                                    FechaOfFin = dp.DueDate.HasValue ? dp.DueDate.Value.ToString("dd/MM/yyyy") : string.Empty,
